Pull transaction output parsing logic out of abstractUtxoCoin
